+++ HISTORY.md
@@ -41,6 +41,7 @@
 ### 2025-11-27
 * gui: refined IDE layout for global magnification zoom levels
 * gui: refined IDE "Magnification" dialog for continuous scaling
+* gui: returned "magnification" item to the settings menu
 
 ### 2025-11-26
 * gui: tweaked global zoom mechanism